Remove create_upstream_oidc_token wrapper
SetupTest now adds nothing of value and tests can call the
underlying function directly, with a slight tweak to how the kid
default is handled.
